% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-package.R
\docType{package}
\name{ollama-package}
\alias{ollama}
\alias{ollama-package}
\title{ollama: Interact with the ollama API}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calderonsamuel/ollama}
  \item \url{https://calderonsamuel.github.io/ollama/}
  \item Report bugs at \url{https://github.com/calderonsamuel/ollama/issues}
}

}
\author{
\strong{Maintainer}: Samuel Calderon \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
