% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{ollama_copy_model}
\alias{ollama_copy_model}
\title{Copy a Model}
\usage{
ollama_copy_model(source, destination)
}
\arguments{
\item{source}{Name of the existing model to copy}

\item{destination}{Name of the new model. Must be different from \code{source}.}
}
\value{
None
}
\description{
Copy a model. Creates a model with another name from an existing model.
}
\examples{
\dontrun{
ollama_copy_model("llama2", "llama2_backup")
}

}
