% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{ollama_delete_model}
\alias{ollama_delete_model}
\title{Delete a Model}
\usage{
ollama_delete_model(name)
}
\arguments{
\item{name}{Model name to delete}
}
\value{
None
}
\description{
Delete a model and its data.
}
\examples{
\dontrun{
ollama_delete_model("llama2")
}

}
