% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama_get_library.R
\name{ollama_get_library}
\alias{ollama_get_library}
\title{Retrieve information from the Ollama AI library.}
\usage{
ollama_get_library()
}
\value{
A tibble with information about models in the Ollama AI library.
}
\description{
This function scrapes the Ollama library page, retrieves information
about available models, and returns a tibble with details such as model name,
description, number of pulls, number of tags, and last update.
}
\examples{
\dontrun{
ollama_get_library()
}

}
