% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama_is_available.R
\name{ollama_is_available}
\alias{ollama_is_available}
\title{Check if Ollama is running}
\usage{
ollama_is_available(verbose = FALSE)
}
\arguments{
\item{verbose}{Display status message?}
}
\value{
A logical value indicating if Ollama is running or not, invisible.
}
\description{
Check if Ollama is running
}
\examples{
ollama_is_available()

}
