% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{ollama_list}
\alias{ollama_list}
\title{List Local Models}
\usage{
ollama_list()
}
\value{
A list with fields \code{name}, \code{modified_at}, and \code{size} for each model.
}
\description{
List models that are available locally.
}
\examples{
\dontrun{
ollama_list()
}

}
