% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{ollama_pull_model}
\alias{ollama_pull_model}
\title{Pull a Model}
\usage{
ollama_pull_model(model, stream = TRUE)
}
\arguments{
\item{stream}{(optional) if FALSE the response will be returned as a single
response object, rather than a stream of objects}

\item{name}{Name of the model to pull}

\item{insecure}{(optional) allow insecure connections to the library.
Only use this if you are pulling from your own library during development.}
}
\value{
None
}
\description{
Download a model from the ollama library.
Cancelled pulls are resumed from where they left off,
and multiple calls will share the same download progress.
}
\examples{
\dontrun{
ollama_pull_model("mistral")
ollama_pull_model("mistral", stream = FALSE)
}
}
